Debug and resolve bugs
Change validation statement for edollar format and a range to prevent sql error.
added if error statement for the entire logical validation so that it will only proceed with logical if the normal validation do not have error to prevent error. debugged and resolved edollar total amount validation. Change all DAO that check if data exist in database using count. Update minutes.

Co-Authored-By: triciatll <triciatll@users.noreply.github.com>