Add bid validations and login page changes
Made change to addBid.php to check if the course selected is not in cart/bid database.
Made changes to Bid.php to make different set of validation for criteria (addbid, cart, bootstrap)
BidDAO to check if bid data exist where userid, courseid and sectionid equals to the search requirement.
Login.php to remove the login instructions that is misleading.
Change placeBidsProcessing.php to process the validation and print error. Clear session afterwards.(Can be further improved)

Co-Authored-By: triciatll <triciatll@users.noreply.github.com>